Task smallwrite (#26)
* Refine shared chunk small write design

* Implement shared small-object chunk writer

* Prepare small-write chunk replacements

* Document shared small-object write path

* Retire completed small-object writer backlog

* Harden small-write pipeline boundaries

* Tune shared small-write pipeline elasticity

* Make chunk writes end-to-end tested

* Repair shared mirror writes in place

* Document small-write replica repair

* Design mirror-to-EC conversion

* Define persistent chunk task contracts

* Add persistent chunk task execution

* Add fenced mirror range replacement

* Add incremental EC parity encoding

* Integrate synchronous EC with small writes

* Run durable mirror conversion tasks

* Harden conversion retry cleanup

* Recover conversion tasks across restart

* Cover active append and batch conversion

* Close conversion recovery races

* Keep foreground EC conversion nonblocking

* Cover conversion deletion recovery

* Document chunk task conversion

* Add unified chunk object reads

* Document chunk object reads

* Add durable chunk read repair

* Document chunk read recovery

* Add resilient large object writes

* Document resilient large object writes

* Add chunk IO read and small write benchmarks

* Document chunk IO benchmark workloads

* Improve chunk IO reliability and regression coverage

* Remove small-write batching delay

* Send only written bytes in small-write mirror RPCs

* Batch strip allocation across chunk writers

* Reserve mirror groups and publish parity in place

* Close reservation recovery and admission gaps

* Remove completed reservation planning artifacts

* Raise small-write EC/mirror sentinel threshold to 80%

* Rewrite latency histograms to HDR and prune internal metrics

* Design unified service configuration

* Unify file-backed service configuration

* Document unified service configuration

* Close unified configuration requirement

* Remove completed small-write and EC working plans

* Update Linux suite timing to 2026-09-10

* Fix flaky kv-ops E2E: gate delayed auto-scan on current toggle and reset cluster in beforeAll

+++ app/crowdb-cli/src/commands/bench/result.rs
@@ -72,7 +72,7 @@ impl From<HistogramSnapshot> for OpStats {
         // LatencyHistogram records in ns; convert to µs for the JSON report.
         Self {
             latency_us: LatencyUs {
-                avg: s.avg / 1000,
+                avg: s.avg / 1000.0,
                 p50: s.p50 / 1000,
                 p99: s.p99 / 1000,
             },
@@ -86,7 +86,7 @@ impl From<HistogramSnapshot> for OpStats {
 #[derive(Debug, Serialize)]
 pub struct LatencyUs {
     #[serde(rename = "avg_us")]
-    pub avg: u64,
+    pub avg: f64,
     #[serde(rename = "p50_us")]
     pub p50: u64,
     #[serde(rename = "p99_us")]
@@ -122,16 +122,16 @@ pub struct ServerMetrics {
 #[derive(Debug, Default, Serialize)]
 #[allow(clippy::struct_field_names)]
 pub struct ServerRpcLatency {
-    pub put_avg_us: u64,
+    pub put_avg_us: f64,
     pub put_p50_us: u64,
     pub put_p99_us: u64,
-    pub get_avg_us: u64,
+    pub get_avg_us: f64,
     pub get_p50_us: u64,
     pub get_p99_us: u64,
-    pub scan_avg_us: u64,
+    pub scan_avg_us: f64,
     pub scan_p50_us: u64,
     pub scan_p99_us: u64,
-    pub delete_avg_us: u64,
+    pub delete_avg_us: f64,
     pub delete_p50_us: u64,
     pub delete_p99_us: u64,
 }
